Refactor truncate! to return truncation error from each bond
This commit refactors the original implemenation. `truncate!` now
expects the user to pass a pointer to a vector of floats with as
many elements as there are bonds in the MPS. It will then store the
truncation error of each bond in the vector.

The corresponding test was updated. The package was re-tested
with the same results described in the PR (75165 passing and
33 broken for total of 75198 tests).

The docstring of `truncate!` was updated to reflect the new behavior.